tests/wpt/css-tests/css-flexbox-1_dev/xhtml1/flexbox_computedstyle_flex-direction-column.xht

<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
<html xmlns="http://www.w3.org/1999/xhtml"><head><title>flexbox | computed style | flex-direction: column</title>
<link href="http://opera.com" rel="author" title="Opera Software" />
<link href="http://www.w3.org/TR/css-flexbox-1/#flex-direction" rel="help" />
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<meta content="dom" name="flags" />
<style>
body {
	color: red;
	flex-direction: column;
}
</style>
</head><body><div id="log"></div>
<script>
test(function() {
    var body = document.body;

    assert_equals(getComputedStyle(body).getPropertyValue("flex-direction"),
	          "column");
});
</script>
</body></html>
